+++ programs/revenue-distribution/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- scaffold rev-distr for integration harvesting (#1057)
 - register rewards integrations (#113)
 
 ## [v0.3.4]